requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
PyJWT==2.8.0
pyotp==2.8.0
qrcode==7.4.2
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.0 